requirements.txt
img2pdf
tqdm
beautifulsoup4
pillow
requests
aiohttp